fix: Don't define terse SCSV schema parser unles Python >= 3.12
It requires `itertools.batched` which was added in Python 3.12.
Also removes the import of `pydrex.io` in `logging`, to avoid cyclical
import errors.
