+++ tests/test_vortex_2d.py
@@ -437,7 +437,7 @@ class TestDiffusionCreep:
         optional_logging = cl.nullcontext()
         if outdir is not None:
             out_basepath = f"{outdir}/{SUBDIR}/{self.class_id}_olA"
-            optional_logging = _log.logfile_enable(f"{out_basepath}.log")
+            optional_logging = _io.logfile_enable(f"{out_basepath}.log")
 
         assert_each_list = [
             get_assert_each(i) for i, _ in enumerate(orientations_init_y)
